Excercicis fets. Començat Ex. Video 1

+++ Exercicis/Conceptes basics - Mis Funciones/LongCalculoFact.java
@@ -7,14 +7,14 @@ public class IntCalculoFact {
 
     public static void main(String[] args) {
         
-        double n, fact = 1;
+        long n, fact = 1;
 
         Scanner scanner = new Scanner(System.in);
         System.out.println("Introduce un numero");
         n = scanner.nextLong();
         System.out.println(fact);
 
-        for (double i = n; i > 1; i--) {
+        for (long i = n; i > 1; i--) {
 
             fact = fact*i;
